project_playwright/requirements.txt
allure-pytest         
allure-python-commons 
attrs                 
colorama              
execnet              
greenlet             
iniconfig            
packaging            
pip                   
playwright            
pluggy               
pyee                  
Pygments              
pytest                
pytest-asyncio       
pytest-rerunfailures  
pytest-xdist          
PyYAML                
setuptools            
typing_extensions     